+++ src/controllers/CsrfTokenController.php
@@ -4,6 +4,9 @@ namespace servd\AssetStorage\controllers;
 
 use Craft;
 use craft\web\Controller;
+use craft\helpers\App;
+use servd\AssetStorage\Plugin;
+use yii\web\NotFoundHttpException;
 
 class CsrfTokenController extends Controller
 {
@@ -11,6 +14,11 @@ class CsrfTokenController extends Controller
 
     public function actionGetToken()
     {
+        $settings = Plugin::$plugin->getSettings();
+        if (!App::env('SERVD_CACHE_ENABLED') || !$settings->injectCors) {
+            throw new NotFoundHttpException();
+        }
+
         $req = Craft::$app->getRequest();
 
         return $this->asJson([