+++ src/controllers/DynamicContentController.php
@@ -3,8 +3,11 @@
 namespace servd\AssetStorage\controllers;
 
 use Craft;
+use craft\helpers\App;
 use craft\web\Controller;
 use craft\web\Request;
+use servd\AssetStorage\Plugin;
+use yii\web\NotFoundHttpException;
 
 class DynamicContentController extends Controller
 {
@@ -13,6 +16,11 @@ class DynamicContentController extends Controller
 
     public function actionGetContent()
     {
+        $settings = Plugin::$plugin->getSettings();
+        if (!App::env('SERVD_CACHE_ENABLED') || $settings->disableDynamic) {
+            throw new NotFoundHttpException();
+        }
+
         /** @var Request $req */
         $req = Craft::$app->getRequest();
         if ($req->getIsCpRequest()) {