+++ src/controllers/DynamicContentController.php
@@ -65,6 +65,9 @@ class DynamicContentController extends Controller
             return $this->asJson($response);
         } else {
             //ESI can only use get requests and only contain a single block
+            if (App::env('SERVD_ESI_ENABLED') !== 'true') {
+                throw new NotFoundHttpException();
+            }
 
             //Make sure the request has a blocks query param
             $blocks = $req->getQueryParam('blocks');