+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Updated
 
 - Updated DynamicContentController#actionGetContent to throw a NotFoundHttpException if static caching or disableDynamic plugin setting is disabled
+- Updated DynamicContentController#actionGetContent ESI path to throw a NotFoundHttpException if ESI is disabled
 - Updated CsrfTokenController#actionGetToken to throw a NotFoundHttpException if static caching or injectCors plugin setting is disabled
 
 ## 3.7.5.1 - 2026-03-24